requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyspark>=3.5.0
kafka-python>=2.0.2
cassandra-driver>=3.28.0

